% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_flynn.R
\name{theme_flynn}
\alias{theme_flynn}
\title{Custom ggplot2 theme}
\usage{
theme_flynn(basesize)
}
\arguments{
\item{basesize}{The size of the font to be used}
}
\value{
Custom ggplot2 theme
}
\description{
This function creates a custom ggplot2 theme
}
\author{
Michael E. Flynn
}
