% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-flynn-map.R
\name{theme_flynn_map}
\alias{theme_flynn_map}
\title{Custom ggplot2 theme}
\usage{
theme_flynn_map(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This function creates a custom ggplot2 theme
}
\author{
Michael E. Flynn
}
